refactor:step2 Split evaluation.ts into smaller modules to fix lint violations
- Extract data loading logic to dataLoader.ts
- Extract evaluation runner logic to evaluationRunner.ts
- Extract metrics calculation to metricsCalculator.ts
- Extract results saving logic to resultsSaver.ts
- Extract validation logic to validator.ts
- Extract case preparation logic to casesPreparer.ts
- Add orchestrator.ts to coordinate data loading and evaluation
- Reduce main evaluation.ts file to under 40 lines
- Update test imports to match new file structure

This refactoring addresses ESLint max-lines-per-function and max-statements violations
by separating concerns into focused modules.
